requirements.txt
asgiref==3.5.2
Django==4.1.3
psycopg2==2.9.5
python-decouple==3.6
sqlparse==0.4.3
tzdata==2022.6
djangorestframework==3.14.0
drf-yasg==1.21.4
django-extensions==3.2.1
ipython==8.6.0
celery==5.2.7
djangorestframework-simplejwt==5.2.2




